(unw_addr_space_t.shared_object_removals): Adjust type to 	"unsigned long long" to match glibc.
(Logical change 1.161)

+++ include/tdep-ia64.h
@@ -101,7 +101,7 @@ struct unw_addr_space
     unw_word_t dyn_generation;
     unw_word_t dyn_info_list_addr;	/* (cached) dyn_info_list_addr */
 #ifndef UNW_REMOTE_ONLY
-    int shared_object_removals;
+    unsigned long long shared_object_removals;
 #endif
 
     struct ia64_script_cache global_cache;
